refactor(master): split traversability into traversability, capability, and protocol-limits validators
