+++ vda5050_core/include/vda5050_core/validation/capability_validator.hpp
@@ -16,22 +16,26 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_CORE__MASTER__VALIDATION__TRAVERSABILITY_VALIDATOR_HPP_
-#define VDA5050_CORE__MASTER__VALIDATION__TRAVERSABILITY_VALIDATOR_HPP_
+#ifndef VDA5050_CORE__MASTER__VALIDATION__CAPABILITY_VALIDATOR_HPP_
+#define VDA5050_CORE__MASTER__VALIDATION__CAPABILITY_VALIDATOR_HPP_
 
 #include "vda5050_core/master/validation/pre_send_validator.hpp"
 #include "vda5050_core/order_utils/validation_result.hpp"
+#include "vda5050_core/types/instant_actions.hpp"
 #include "vda5050_core/types/order.hpp"
 
 namespace vda5050_core::master {
 
-/// \brief Can this AGV navigate the order's route? Validates first-node
-/// reachability (state-driven) and, when a layout is loaded, that every
-/// node/edge exists in it with matching map_id and directed endpoints. The
-/// graph-integrity check is skipped when no layout is cached.
-vda5050_core::order_utils::ValidationResult validate_traversability(
+/// \brief Can this AGV perform the message's actions? Checks every action
+/// against the factsheet (action type supported, required scope, blocking
+/// type, declared/required parameters). Skips when no factsheet is cached.
+vda5050_core::order_utils::ValidationResult validate_capability(
   const PreSendContext& ctx, const vda5050_core::types::Order& order);
+
+vda5050_core::order_utils::ValidationResult validate_capability(
+  const PreSendContext& ctx,
+  const vda5050_core::types::InstantActions& actions);
 
 }  // namespace vda5050_core::master
 
-#endif  // VDA5050_CORE__MASTER__VALIDATION__TRAVERSABILITY_VALIDATOR_HPP_
+#endif  // VDA5050_CORE__MASTER__VALIDATION__CAPABILITY_VALIDATOR_HPP_
