+++ vda5050_core/test/master/test_unit_traversability_validator.cpp
@@ -65,32 +65,11 @@ namespace vda5050_core::master::test {
 
 namespace {
 
-// Build a Factsheet that supports the typical happy-path Order: generous
-// limits, one supported action_type "pick" available on NODE/EDGE/INSTANT
-// scopes with a "loadId" parameter key.
+// Factsheet is irrelevant to traversability (which is state + layout driven),
+// but PreSendContext requires one; this is a minimal stand-in.
 vda5050_core::types::Factsheet make_factsheet()
 {
-  vda5050_core::types::Factsheet fs;
-  fs.protocol_limits.max_array_lens.order_nodes = 100;
-  fs.protocol_limits.max_array_lens.order_edges = 100;
-  fs.protocol_limits.max_array_lens.node_actions = 10;
-  fs.protocol_limits.max_array_lens.edge_actions = 10;
-  fs.protocol_limits.max_array_lens.actions_actions_parameters = 10;
-  fs.protocol_limits.max_array_lens.instant_actions = 10;
-
-  vda5050_core::types::AGVAction supported;
-  supported.action_type = "pick";
-  supported.action_scopes = {
-    vda5050_core::types::ActionScope::NODE,
-    vda5050_core::types::ActionScope::EDGE,
-    vda5050_core::types::ActionScope::INSTANT};
-  vda5050_core::types::ActionParameterFactsheet param;
-  param.key = "loadId";
-  param.is_optional = true;
-  supported.action_parameters =
-    std::vector<vda5050_core::types::ActionParameterFactsheet>{param};
-  fs.protocol_features.agv_actions = {supported};
-  return fs;
+  return vda5050_core::types::Factsheet{};
 }
 
 // State where AGV is on node "N0" — trivial reachability for an order whose
@@ -127,15 +106,6 @@ vda5050_core::types::Order make_minimal_order()
   n.released = true;
   o.nodes.push_back(n);
   return o;
-}
-
-vda5050_core::types::Action make_pick_action()
-{
-  vda5050_core::types::Action a;
-  a.action_id = "A1";
-  a.action_type = "pick";
-  a.blocking_type = vda5050_core::types::BlockingType::NONE;
-  return a;
 }
 
 }  // namespace
@@ -248,303 +218,6 @@ TEST(TraversabilityValidatorTest, OrderRejectFirstNodeOnDifferentMap)
   EXPECT_FALSE(static_cast<bool>(res));
   EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
   EXPECT_TRUE(AnyErrorMentions(res, "different map"));
-}
-
-// ============================================================================
-// Order — action capability rejections
-// ============================================================================
-
-TEST(TraversabilityValidatorTest, OrderRejectActionTypeNotInFactsheet)
-{
-  auto ctx = make_ctx(make_state_on_node("N0"), make_factsheet());
-  auto order = make_minimal_order();
-  vda5050_core::types::Action unsupported;
-  unsupported.action_id = "A1";
-  unsupported.action_type = "fly";  // not in factsheet
-  unsupported.blocking_type = vda5050_core::types::BlockingType::NONE;
-  order.nodes.front().actions = {unsupported};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "fly"));
-}
-
-TEST(TraversabilityValidatorTest, OrderRejectActionScopeNotNode)
-{
-  // Build factsheet where "pick" is INSTANT-only — not allowed on Nodes.
-  auto fs = make_factsheet();
-  fs.protocol_features.agv_actions[0].action_scopes = {
-    vda5050_core::types::ActionScope::INSTANT};
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  order.nodes.front().actions = {make_pick_action()};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "scope"));
-}
-
-TEST(TraversabilityValidatorTest, OrderRejectBlockingTypeNotSupported)
-{
-  // Factsheet declares "pick" supports HARD only; order requests SOFT.
-  auto fs = make_factsheet();
-  fs.protocol_features.agv_actions[0].blocking_types = {
-    vda5050_core::types::BlockingType::HARD};
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  auto a = make_pick_action();
-  a.blocking_type = vda5050_core::types::BlockingType::SOFT;
-  order.nodes.front().actions = {a};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "blocking_type"));
-}
-
-TEST(TraversabilityValidatorTest, OrderAcceptBlockingTypeSupported)
-{
-  // Factsheet declares "pick" supports NONE and HARD; order requests NONE.
-  auto fs = make_factsheet();
-  fs.protocol_features.agv_actions[0].blocking_types = {
-    vda5050_core::types::BlockingType::NONE,
-    vda5050_core::types::BlockingType::HARD};
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  order.nodes.front().actions = {make_pick_action()};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_TRUE(static_cast<bool>(res));
-}
-
-TEST(
-  TraversabilityValidatorTest, OrderSkipsBlockingTypeCheckWhenFactsheetSilent)
-{
-  // make_factsheet leaves blocking_types unset — any blocking_type accepted.
-  auto ctx = make_ctx(make_state_on_node("N0"), make_factsheet());
-  auto order = make_minimal_order();
-  auto a = make_pick_action();
-  a.blocking_type = vda5050_core::types::BlockingType::SOFT;
-  order.nodes.front().actions = {a};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_TRUE(static_cast<bool>(res));
-}
-
-TEST(TraversabilityValidatorTest, OrderRejectActionParameterKeyNotInFactsheet)
-{
-  auto ctx = make_ctx(make_state_on_node("N0"), make_factsheet());
-  auto order = make_minimal_order();
-  auto a = make_pick_action();
-  vda5050_core::types::ActionParameter p;
-  p.key = "weirdKey";  // not declared by factsheet
-  p.value = "x";
-  a.action_parameters = std::vector<vda5050_core::types::ActionParameter>{p};
-  order.nodes.front().actions = {a};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "weirdKey"));
-}
-
-TEST(TraversabilityValidatorTest, OrderRejectMissingRequiredActionParam)
-{
-  auto fs = make_factsheet();
-  fs.protocol_features.agv_actions[0].action_parameters->front().is_optional =
-    false;  // loadId now required
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  order.nodes.front().actions = {make_pick_action()};  // no loadId supplied
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "missing required parameter"));
-}
-
-TEST(TraversabilityValidatorTest, OrderAcceptsRequiredActionParamPresent)
-{
-  auto fs = make_factsheet();
-  fs.protocol_features.agv_actions[0].action_parameters->front().is_optional =
-    false;
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  auto a = make_pick_action();
-  vda5050_core::types::ActionParameter p;
-  p.key = "loadId";
-  p.value = "L1";
-  a.action_parameters = std::vector<vda5050_core::types::ActionParameter>{p};
-  order.nodes.front().actions = {a};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_TRUE(static_cast<bool>(res));
-}
-
-// ============================================================================
-// Order — array-size limit rejections
-// ============================================================================
-
-TEST(TraversabilityValidatorTest, OrderRejectExceedsMaxNodes)
-{
-  auto fs = make_factsheet();
-  fs.protocol_limits.max_array_lens.order_nodes = 1;  // tight limit
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  vda5050_core::types::Node n2;
-  n2.node_id = "N1";
-  n2.sequence_id = 2;
-  n2.released = true;
-  order.nodes.push_back(n2);
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "order_nodes"));
-}
-
-TEST(TraversabilityValidatorTest, OrderRejectExceedsMaxNodeActions)
-{
-  auto fs = make_factsheet();
-  fs.protocol_limits.max_array_lens.node_actions = 1;
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  order.nodes.front().actions = {make_pick_action(), make_pick_action()};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "node_actions"));
-}
-
-// ============================================================================
-// InstantActions — capability + limit
-// ============================================================================
-
-TEST(TraversabilityValidatorTest, InstantActionsHappyPath)
-{
-  auto ctx = make_ctx(make_state_on_node("N0"), make_factsheet());
-  vda5050_core::types::InstantActions ia;
-  ia.actions = {make_pick_action()};
-  auto res = validate_traversability(ctx, ia);
-  EXPECT_TRUE(static_cast<bool>(res));
-}
-
-TEST(TraversabilityValidatorTest, InstantActionsRejectActionTypeNotInFactsheet)
-{
-  auto ctx = make_ctx(make_state_on_node("N0"), make_factsheet());
-  vda5050_core::types::InstantActions ia;
-  vda5050_core::types::Action unsupported;
-  unsupported.action_id = "A1";
-  unsupported.action_type = "teleport";
-  unsupported.blocking_type = vda5050_core::types::BlockingType::NONE;
-  ia.actions = {unsupported};
-  auto res = validate_traversability(ctx, ia);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "teleport"));
-}
-
-TEST(TraversabilityValidatorTest, InstantActionsRejectScopeNotInstant)
-{
-  auto fs = make_factsheet();
-  fs.protocol_features.agv_actions[0].action_scopes = {
-    vda5050_core::types::ActionScope::NODE};  // NODE only, NOT INSTANT
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-  vda5050_core::types::InstantActions ia;
-  ia.actions = {make_pick_action()};
-  auto res = validate_traversability(ctx, ia);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "scope"));
-}
-
-TEST(TraversabilityValidatorTest, InstantActionsRejectExceedsMaxLength)
-{
-  auto fs = make_factsheet();
-  fs.protocol_limits.max_array_lens.instant_actions = 1;
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-  vda5050_core::types::InstantActions ia;
-  ia.actions = {make_pick_action(), make_pick_action()};
-  auto res = validate_traversability(ctx, ia);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "instant_actions"));
-}
-
-// ============================================================================
-// Skip-when-factsheet-missing
-// ============================================================================
-
-TEST(TraversabilityValidatorTest, OrderSkipsCapabilityChecksWhenNoFactsheet)
-{
-  // No factsheet → skip capability + limit checks. Reachability still runs;
-  // AGV is on first node so it passes.
-  PreSendContext ctx{
-    vda5050_core::types::ConnectionState::ONLINE, make_state_on_node("N0"),
-    std::nullopt, AGVState::AVAILABLE, nullptr};
-  auto order = make_minimal_order();
-  auto res = validate_traversability(ctx, order);
-  EXPECT_TRUE(static_cast<bool>(res));
-}
-
-TEST(
-  TraversabilityValidatorTest,
-  InstantActionsSkipsCapabilityChecksWhenNoFactsheet)
-{
-  // No factsheet → no capability check on InstantActions; passes silently.
-  PreSendContext ctx{
-    vda5050_core::types::ConnectionState::ONLINE, make_state_on_node("N0"),
-    std::nullopt, AGVState::AVAILABLE, nullptr};
-  vda5050_core::types::InstantActions ia;
-  vda5050_core::types::Action a;
-  a.action_id = "A1";
-  a.action_type = "anything_goes";
-  a.blocking_type = vda5050_core::types::BlockingType::NONE;
-  ia.actions = {a};
-  auto res = validate_traversability(ctx, ia);
-  EXPECT_TRUE(static_cast<bool>(res));
-}
-
-// ============================================================================
-// Multi-error accumulation
-// ============================================================================
-
-TEST(TraversabilityValidatorTest, MultipleErrorsAccumulate)
-{
-  auto fs = make_factsheet();
-  fs.protocol_limits.max_array_lens.order_nodes = 1;  // tight limit
-  auto ctx = make_ctx(make_state_on_node("N0"), fs);
-
-  auto order = make_minimal_order();
-  // Add a second node to exceed limit
-  vda5050_core::types::Node n2;
-  n2.node_id = "N1";
-  n2.sequence_id = 2;
-  n2.released = true;
-  order.nodes.push_back(n2);
-  // Add an unsupported action type
-  vda5050_core::types::Action unsupported;
-  unsupported.action_id = "A1";
-  unsupported.action_type = "fly";
-  unsupported.blocking_type = vda5050_core::types::BlockingType::NONE;
-  order.nodes.front().actions = {unsupported};
-
-  auto res = validate_traversability(ctx, order);
-  EXPECT_FALSE(static_cast<bool>(res));
-  EXPECT_GE(res.errors.size(), 2u);
-  EXPECT_TRUE(AllErrorsHaveTraversabilityType(res));
-  EXPECT_TRUE(AnyErrorMentions(res, "order_nodes"));
-  EXPECT_TRUE(AnyErrorMentions(res, "fly"));
 }
 
 // ============================================================================
